Show funny loading messages on all loading screens
Extract shared LoadingContent composable that combines a spinner with a random Easter egg message from generic_loading_labels. Previously only AppDetailsScreen showed these messages; now all five loading screens use the shared component.

Also fix permissions list showing blank screen instead of loading indicator when state is initially null.
